+++ tailwind.config.cjs
@@ -11,15 +11,6 @@ const config = {
       xl: "1440px",
       xxl: "2160px",
     },
-    colors: {
-      background: "#FFFFFF",
-      "accent-1": "#FF5252",
-      "light-accent-1": "hsl(0, 100%, 66%, 0.6)",
-      "accent-2": "#F45420",
-      secondary: "#EAEAEA",
-      text: "#121212",
-      transparent: "hsl(0, 0%, 0%, 0.0)",
-    },
     fontFamily: {
       poppins: ["Poppins", "sans-serif"],
     },
@@ -27,7 +18,17 @@ const config = {
       default: "2px 4px 32px rgba(155, 49, 49, 0.1)",
       "default-solid": "8px 8px 0px #FF5252",
     },
-    extend: {},
+    extend: {
+      colors: {
+        background: "#FFFFFF",
+        "accent-1": "#FF5252",
+        "light-accent-1": "hsl(0, 100%, 66%, 0.6)",
+        "accent-2": "#F45420",
+        secondary: "#EAEAEA",
+        text: "#121212",
+        transparent: "hsl(0, 0%, 0%, 0.0)",
+      },
+    },
   },
   plugins: [],
 };
